Move XlaTensor to a separate dir and rename XlaTensorOwned to XlaTensorStringBuf
